layout changes

+++ zamg2jpg_arcgispro.py
@@ -4,9 +4,12 @@
 Created on Sun Jan 31 11:44:33 2020
 
 LV: 2020W 290065-1 Grundlagen der Softwareentwicklung
+ausführbar in jedem leeren Projekt,
+optionale Parameter:
+0   Layout File
 """
 # imports
-import requests, os, datetime
+import requests, os, datetime, time
 
 # download of daily (last 24h) earthquake data from ZAMG
 # alternative file types: csv
@@ -35,27 +38,32 @@ def try2float(mag):
 
 arcpy.management.CalculateField('quakes', 'Magnitude', expression, 'PYTHON3', code_block)
 
-# expression für Point size: exp("Magnitude")
-# ? Auf Österreich zentrieren, weil NÖ Daten?
-
-# Importing Layout first to work around broken Data Source Path
-url = "https://raw.githubusercontent.com/ryepenchi/zamg2jpeg/main/Layout.pagx"
-r = requests.get(url)
-with open("Layout.pagx", "wb") as f:
-    f.write(r.content)
-p.importDocument(os.path.join(p.homeFolder, "Layout.pagx"))
+# Check whether to use specific or local Layout File, download if none exists
+# Parameter 0 -> localLayout
+if arcpy.GetParameterAsText(0):
+    arcpy.AddMessage("Using specified Layout File...")
+    layout = arcpy.GetParameter(0)
+elif not os.path.exists("Layout.pagx"):
+    arcpy.AddMessage("Downloading Layout File...")
+    url = "https://raw.githubusercontent.com/ryepenchi/zamg2jpeg/main/Layout.pagx"
+    r = requests.get(url)
+    with open("Layout.pagx", "wb") as f:
+        f.write(r.content)
+    layout = os.path.join(p.homeFolder, "Layout.pagx")
+else:
+    arcpy.AddMessage("Using local Layout file...")
+    layout = os.path.join(p.homeFolder, "Layout.pagx")
+p.importDocument(layout)
 lyt = p.listLayouts("Layout")[0]
 
 # Defining the Layer
-# m = p.listMaps('Map')[0]
-
-# m.addBasemap("Hellgrauer Hintergrund")
-# m.addDataFromPath(os.path.join(p.defaultGeodatabase, "quakes"))
+# m.addBasemap("Hellgrauer Hintergrund") # bereits im Layout File definiert
+# m.addDataFromPath(os.path.join(p.defaultGeodatabase, "quakes"))# anscheinend nicht dasselbe wie hinzufügen zum Layout Element
+# lyr = m.listLayers('quakes')[0]
 lyt.listElements("MAPFRAME_ELEMENT")[0].map.addDataFromPath(os.path.join(p.defaultGeodatabase, "quakes"))
-
-#lyr = m.listLayers('quakes')[0]
 lyr = lyt.listElements("MAPFRAME_ELEMENT")[0].map.listLayers("quakes")[0]
 
+# Einige manuelle Einstellungen der Symbology
 sym = lyr.symbology
 sym.updateRenderer('GraduatedSymbolsRenderer')
 sym.renderer.classificationField = 'Magnitude'
@@ -63,22 +71,22 @@ sym.renderer.normalizationField = None
 sym.renderer.colorRamp = p.listColorRamps("Yellow to Red")[0]
 sym.classBreakValues = ['1.5', '2.0', '3.0', '4.0', '5.0', '6.0']
 sym.renderer.breakCount = len(sym.classBreakValues)
-#update Break value labels
+# update Break value labels
 for brk in sym.renderer.classBreaks:
     brk.label = brk.label.rstrip("0")
 lyr.symbology = sym
 lyr.transparency = 30
 
-# Making the Legend
-timeformat = "%d.%m.%Y"
+# Adding timestamp
 now = datetime.datetime.now()
-# yesterday = now - datetime.timedelta(days = 1)
+timeformat = "%d.%m.%Y %H:%M UTC+" + str(time.timezone // -3600)
+# timeformat = "%x %X (UTC +1)"
 now = now.strftime(timeformat)
-# yesterday = yesterday.strftime(timeformat)
 
-lyt.listElements("LEGEND_ELEMENT")[0].showTitle = True
-lyt.listElements("LEGEND_ELEMENT")[0].title = now
+# lyt.listElements("LEGEND_ELEMENT")[0].showTitle = True
+for t in lyt.listElements("TEXT_ELEMENT"):
+    if t.name == "date":
+        t.text = "Zeitpunkt: " + now
 
 # Exporting to jpg
-
 lyt.exportToJPEG("exportmap.jpg")
